# fern/apis/prod/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: Superagent
  description: The Open Source AI Assistant Framework & API
  version: 0.2.13
servers:
  - url: https://api.beta.superagent.sh
paths:
...
  /api/v1/workflows:
    get:
      tags:
        - Workflow
      summary: List
      description: List all workflows
      operationId: list_api_v1_workflows_get
      parameters:
        - required: false
          schema:
            title: Skip
            type: integer
            default: 0
          name: skip
          in: query
        - required: false
          schema:
            title: Take
            type: integer
            default: 50
          name: take
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
    post:
      tags:
        - Workflow
      summary: Create
      description: Create a new workflow
      operationId: create_api_v1_workflows_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__models__request__Workflow'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__Workflow'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /api/v1/workflows/{workflow_id}:
    get:
      tags:
        - Workflow
      summary: Get
      description: Get a single workflow
      operationId: get_api_v1_workflows__workflow_id__get
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__Workflow'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
    delete:
      tags:
        - Workflow
      summary: Delete
      description: Delete a specific workflow
      operationId: delete_api_v1_workflows__workflow_id__delete
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
    patch:
      tags:
        - Workflow
      summary: Update
      description: Patch a workflow
      operationId: update_api_v1_workflows__workflow_id__patch
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__models__request__Workflow'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__Workflow'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /api/v1/workflows/{workflow_id}/invoke:
    post:
      tags:
        - Workflow
      summary: Invoke
      description: Invoke a specific workflow
      operationId: invoke_api_v1_workflows__workflow_id__invoke_post
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowInvoke'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /api/v1/workflows/{workflow_id}/steps:
    get:
      tags:
        - Workflow
      summary: List Steps
      description: List all steps of a workflow
      operationId: list_steps_api_v1_workflows__workflow_id__steps_get
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStepList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
    post:
      tags:
        - Workflow
      summary: Add Step
      description: Create a new workflow step
      operationId: add_step_api_v1_workflows__workflow_id__steps_post
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__models__request__WorkflowStep'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__WorkflowStep'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /api/v1/workflows/{workflow_id}/steps/{step_id}:
    delete:
      tags:
        - Workflow
      summary: Delete Step
      description: Delete a specific workflow step
      operationId: delete_step_api_v1_workflows__workflow_id__steps__step_id__delete
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
        - required: true
          schema:
            title: Step Id
            type: string
          name: step_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
    patch:
      tags:
        - Workflow
      summary: Update
      description: Patch a workflow step
      operationId: update_api_v1_workflows__workflow_id__steps__step_id__patch
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
          name: workflow_id
          in: path
        - required: true
          schema:
            title: Step Id
            type: string
          name: step_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__models__request__WorkflowStep'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__WorkflowStep'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
...
components:
  schemas:
...
    AgentType:
      title: AgentType
      enum:
        - SUPERAGENT
        - OPENAI_ASSISTANT
        - LLM
      type: string
      description: An enumeration.
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    LLMModel:
      title: LLMModel
      enum:
        - GPT_3_5_TURBO_16K_0613
        - GPT_3_5_TURBO_0613
        - GPT_3_5_TURBO_1106
        - GPT_3_5_TURBO_0125
        - GPT_4_0613
        - GPT_4_32K_0613
        - GPT_4_1106_PREVIEW
        - GPT_4_TURBO_PREVIEW
        - MISTRAL_7B_INSTRUCT_V01
        - MIXTRAL_8X7B_INSTRUCT_V01
      type: string
      description: An enumeration.
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
    WorkflowConfig:
      title: WorkflowConfig
      required:
        - id
        - createdAt
        - updatedAt
        - workflowId
      type: object
      properties:
        id:
          title: Id
          type: string
        config:
          title: Config
          type: string
          format: json-string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        workflowId:
          title: Workflowid
          type: string
        workflow:
          $ref: '#/components/schemas/prisma__models__Workflow'
        ApiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        apiUserId:
          title: Apiuserid
          type: string
      description: Represents a WorkflowConfig record
    WorkflowInvoke:
      title: WorkflowInvoke
      required:
        - input
        - enableStreaming
      type: object
      properties:
        input:
          title: Input
          type: string
        enableStreaming:
          title: Enablestreaming
          type: boolean
        sessionId:
          title: Sessionid
          type: string
    WorkflowList:
      title: WorkflowList
      required:
        - success
        - total_pages
      type: object
      properties:
        success:
          title: Success
          type: boolean
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Workflow'
        total_pages:
          title: Total Pages
          type: integer
    WorkflowStepList:
      title: WorkflowStepList
      required:
        - success
      type: object
      properties:
        success:
          title: Success
          type: boolean
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__WorkflowStep'
...
    app__models__request__Workflow:
      title: Workflow
      required:
        - name
        - description
      type: object
      properties:
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
    app__models__request__WorkflowStep:
      title: WorkflowStep
      required:
        - order
        - agentId
      type: object
      properties:
        order:
          title: Order
          type: integer
        agentId:
          title: Agentid
          type: string
...
    app__models__response__Workflow:
      title: Workflow
      required:
        - success
      type: object
      properties:
        success:
          title: Success
          type: boolean
        data:
          $ref: '#/components/schemas/prisma__models__Workflow'
    app__models__response__WorkflowStep:
      title: WorkflowStep
      required:
        - success
      type: object
      properties:
        success:
          title: Success
          type: boolean
        data:
          $ref: '#/components/schemas/prisma__models__WorkflowStep'
    prisma__models__Agent:
      title: Agent
      required:
        - id
        - type
        - name
        - description
        - isActive
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        type:
          $ref: '#/components/schemas/AgentType'
        name:
          title: Name
          type: string
        avatar:
          title: Avatar
          type: string
        initialMessage:
          title: Initialmessage
          type: string
        description:
          title: Description
          type: string
        isActive:
          title: Isactive
          type: boolean
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        llms:
          title: Llms
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentLLM'
        llmModel:
          $ref: '#/components/schemas/LLMModel'
        prompt:
          title: Prompt
          type: string
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentDatasource'
        tools:
          title: Tools
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentTool'
        workflowSteps:
          title: Workflowsteps
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__WorkflowStep'
        metadata:
          title: Metadata
          type: string
          format: json-string
      description: Represents a Agent record
...
    prisma__models__ApiUser:
      title: ApiUser
      required:
        - id
        - createdAt
        - updatedAt
      type: object
      properties:
        id:
          title: Id
          type: string
        token:
          title: Token
          type: string
        email:
          title: Email
          type: string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        agents:
          title: Agents
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Agent'
        llms:
          title: Llms
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__LLM'
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Datasource'
        tools:
          title: Tools
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Tool'
        workflows:
          title: Workflows
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Workflow'
        vectorDb:
          title: Vectordb
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__VectorDb'
        workflowConfigs:
          title: Workflowconfigs
          type: array
          items:
            $ref: '#/components/schemas/WorkflowConfig'
        apiKeys:
          title: Apikeys
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__ApiKey'
      description: Represents a ApiUser record
    prisma__models__Datasource:
      title: Datasource
      required:
        - id
        - name
        - type
        - apiUserId
        - createdAt
        - updatedAt
        - status
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        content:
          title: Content
          type: string
        description:
          title: Description
          type: string
        url:
          title: Url
          type: string
        type:
          $ref: '#/components/schemas/DatasourceType'
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        metadata:
          title: Metadata
          type: string
        status:
          $ref: '#/components/schemas/DatasourceStatus'
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentDatasource'
        vectorDb:
          $ref: '#/components/schemas/prisma__models__VectorDb'
        vectorDbId:
          title: Vectordbid
          type: string
      description: Represents a Datasource record
    prisma__models__LLM:
      title: LLM
      required:
        - id
        - provider
        - apiKey
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        provider:
          $ref: '#/components/schemas/LLMProvider'
        apiKey:
          title: Apikey
          type: string
        options:
          title: Options
          type: string
          format: json-string
        agents:
          title: Agents
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentLLM'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
      description: Represents a LLM record
...
    prisma__models__Workflow:
      title: Workflow
      required:
        - id
        - name
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        steps:
          title: Steps
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__WorkflowStep'
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        workflowConfigs:
          title: Workflowconfigs
          type: array
          items:
            $ref: '#/components/schemas/WorkflowConfig'
      description: Represents a Workflow record
    prisma__models__WorkflowStep:
      title: WorkflowStep
      required:
        - id
        - order
        - workflowId
        - createdAt
        - updatedAt
        - agentId
      type: object
      properties:
        id:
          title: Id
          type: string
        order:
          title: Order
          type: integer
        workflowId:
          title: Workflowid
          type: string
        workflow:
          $ref: '#/components/schemas/prisma__models__Workflow'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        input:
          title: Input
          type: string
        output:
          title: Output
          type: string
        agentId:
          title: Agentid
          type: string
        agent:
          $ref: '#/components/schemas/prisma__models__Agent'
      description: Represents a WorkflowStep record
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer